requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.1.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==2.3.2
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.0.5
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
PyJWT==2.10.1
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2024.2
requests==2.32.3
SQLAlchemy==2.0.37
twilio==8.3.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3